usnic: Move missed AV extensions to usdf_ext.c file.
